Move initOptions to /src/lib
